+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * Add confirmation dialogs to dangerous operations
 * Add spinner loader to indicate when page is loading
 * Improve error output when a template fails to render in the GUI
-* Improve error output for bad "name" parameters on POST to /applications
+* Improve error output for bad "name" property on POST to /applications
+* Improve error output for missing "name" or "parameters" property on POST to /applications
 * Add showDisabled query parameter to GET on /templatesets for showing disabled template sets
 * Support PATCH on /applications/{tenant}/{app} endpoint for partial modifications of existing applications
 * Add template set hash (setHash) to application data